requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libretranslatepy==2.1.1
lxml==4.9.3
MarkupSafe==2.1.3
Pillow==10.0.0
pyjokes==0.6.0
PyPDF2==1.26.0
requests==2.31.0
soupsieve==2.4.1
translate==3.6.1
urllib3==2.0.4
Werkzeug==2.3.7
